Move `ContractMetadataKind` into `metadata` module
