+++ crates/contracts/ab-contracts-macros/src/__private.rs
@@ -1,6 +1,7 @@
 pub use ab_contracts_common::env::{Env, MethodContext};
+pub use ab_contracts_common::metadata::ContractMetadataKind;
 pub use ab_contracts_common::methods::{ExternalArgs, MethodFingerprint};
-pub use ab_contracts_common::{Address, ContractError, ContractMetadataKind, ExitCode};
+pub use ab_contracts_common::{Address, ContractError, ExitCode};
 pub use ab_contracts_io_type::trivial_type::TrivialType;
 pub use ab_contracts_io_type::utils::concat_metadata_sources;
 pub use ab_contracts_io_type::{IoType, IoTypeOptional};
